+++ tests/integration_tests/adapters/binance/test_parsing_common.py
@@ -15,12 +15,24 @@
 
 import pytest
 
+from nautilus_trader.adapters.binance.common.parsing.data import parse_bar_ws
+from nautilus_trader.adapters.binance.common.schemas import BinanceCandlestick
 from nautilus_trader.adapters.binance.spot.enums import BinanceSpotOrderType
 from nautilus_trader.adapters.binance.spot.parsing.execution import parse_order_type
+from nautilus_trader.backtest.data.providers import TestInstrumentProvider
+from nautilus_trader.core.datetime import millis_to_nanos
+from nautilus_trader.model.data.bar import BarSpecification
+from nautilus_trader.model.data.bar import BarType
+from nautilus_trader.model.enums import AggregationSource
+from nautilus_trader.model.enums import BarAggregation
 from nautilus_trader.model.enums import OrderType
+from nautilus_trader.model.enums import PriceType
 
 
-class TestBinanceSpotParsing:
+BTCUSDT_BINANCE = TestInstrumentProvider.ethusdt_binance()
+
+
+class TestBinanceCommonParsing:
     @pytest.mark.parametrize(
         "order_type, expected",
         [
@@ -38,3 +50,160 @@ class TestBinanceSpotParsing:
 
         # Assert
         assert result == expected
+
+    @pytest.mark.parametrize(
+        "resolution, expected_type",
+        [
+            [
+                "1m",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(1, BarAggregation.MINUTE, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "3m",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(3, BarAggregation.MINUTE, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "5m",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(5, BarAggregation.MINUTE, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "15m",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(15, BarAggregation.MINUTE, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "30m",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(30, BarAggregation.MINUTE, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "1h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(1, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "2h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(2, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "4h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(4, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "6h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(6, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "8h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(8, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "12h",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(12, BarAggregation.HOUR, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "1d",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(1, BarAggregation.DAY, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "3d",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(3, BarAggregation.DAY, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "1w",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(1, BarAggregation.WEEK, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+            [
+                "1M",
+                BarType(
+                    BTCUSDT_BINANCE.id,
+                    BarSpecification(1, BarAggregation.MONTH, PriceType.LAST),
+                    AggregationSource.EXTERNAL,
+                ),
+            ],
+        ],
+    )
+    def test_parse_parse_bar_ws(self, resolution, expected_type):
+        # Arrange
+        candle = BinanceCandlestick(
+            t=1638747660000,
+            T=1638747719999,
+            s="BTCUSDT",
+            i=resolution,
+            f=100,
+            L=200,
+            o="0.0010",
+            c="0.0020",
+            h="0.0025",
+            l="0.0015",
+            v="1000",
+            n=100,
+            x=False,
+            q="1.0000",
+            V="500",
+            Q="0.500",
+            B="123456",
+        )
+
+        # Act
+        bar = parse_bar_ws(
+            instrument_id=BTCUSDT_BINANCE.id,
+            data=candle,
+            ts_init=millis_to_nanos(1638747720000),
+        )
+
+        # Assert
+        assert bar.type == expected_type