+++ RELEASES.md
@@ -11,14 +11,15 @@ Released on TBD.
 - Replaced `UUID4.value` property with `UUID4.to_str()` method
 
 ### Enhancements
+- Added `WEEK` and `MONTH` bar aggregation options
 - Added `Position.closing_order_id` property
 - Added `tags` param to `Strategy.submit_order`
 - Added optional `check_positon_exists` flag to `Strategy.submit_order`
-- Eliminated all use of `unsafe` Rust and nul byte terminated C strings.
+- Eliminated all use of `unsafe` Rust and nul byte terminated C strings
 - The `bypass_logging` config option will also now bypass the `BacktestEngine` logger
 
 ### Fixes
-None
+- Fix Binance bar resolution parsing
 
 ---